Add two Linux Ouya Controller mappings
These mappings are the same as the Linux Ouya mapping in SDL2's gamecontrollerdb.h, but different GUIDs.
The first one is the from a pre-ordered console+controller and second is the from pre-ordered controllers
I received a few months later (an email said they were "latest controller(s) with the most up-to-date firmware",
so different GUID wasn't surprising).