Add RetroUSB.com SNES RetroPort mapping for Linux and change Windows mapping
Added Linux mapping for RetroUSB.com SNES RetroPort.

Changed existing Windows mapping for the same adapter; A, B, X and Y
buttons were mapped according to the labels on the front of the SNES
controller.  They are now mapped in the same way as other supported
controllers with the same button arrangement (using the XBox 360
button naming) for consistency.