Add controller strings for SFC30
SFC30 controller: http://www.8bitdo.com/sfc30/

The SFC30 controller can present itself in a variety of modes and it offers up
different names in each. This patch captures data for one mode on three
platforms (Mac OS X, Windows, Linux).

(USB mode on Linux and Windows is missing as the button events did not make it
through to SDL's controllermap tool on Fedora 24/Linux 4.5.5 nor Steam Big
Picture mode on Windows. The two Bluetooth modes were indistinguishable on
Windows.)

There exists a similar controller called the SNES30 (And some others) that are
very likely identical except for the name, but I have not verified this yet so
haven't synthesized lines for those controllers until I can.

This patch incorporates the changes made in #84 (thanks @mariuszk911 which are
correct but not comprehensive.  (I could've just supplied the missing ones but
each of these PRs are going to conflict anyway)