Add Switch Joy-Cons
This adds Nintendo Switch Joy-Con support as "Joy-Con (L)" and "Joy-Con (R)".
These are for the Joy-Cons in their horizontal position, so they only have
the x/y/a/b buttons, the left stick (including press), the shoulder buttons,
and the back/start buttons. The Joy-Con's vertical orientation trigger and
button aren't mapped to anything. This maps the back button to whichever of
the +/-/screenshot/home button is on the left, and the start button to
whichever is on the right.