Update Mayflash GameCube Controller mapping on Linux
I have a "Mayflash GameCube Controller Adapter", and I believe the A and X buttons are erroneously swapped in the Linux mapping. Note that this change would bring the Linux map for this device into line with both "Mayflash GameCube Controller" on Linux and the Windows mappings of the same.