Added 10 gamepads on Mac and 2 on Windows
Mac
- Gravis Gamepad Pro
- Switch PowerA Wired Controller
- AxisPad
- Generic Joystick
- Xbox One PowerA Wired Controller
- PS2 Adapter
- Wii U Pro Controller (Mayflash Adapter, HID Mode)
- Xbox 360 Wired Controller
- NES 2-port Adapter
- GameCube Controller Adapter

Windows
- AxisPad
- NES 2-port Adapter

The "NES 2-port Adapter" is mapped A->b and B->a, so that it matches the position of Xbox controllers more. From what I understand that is the convention for the database. The "GameCube Controller Adapter" on Mac is mapped A->a, B->b, X->x, Y->y, because the buttons are positioned kinda weird on GameCube and this is how it already is on Windows in the database.