Update gamecontrollerdb.txt
Adds and reconfigures GUIDs that have overlap as both Sega Mega Drive/Saturn and Super Nintendo/Famicom style gamepads. Variants include a number of devices sold by the brands Retrolink, Retro-bit, 8BitDo, Retro Fighters, Buffalo, Micreal and Ysteck. In each case the Saturn style layout is a superset of the SNES layout (with face buttons reordered so the labels are the same), as such we're defaulting to the Saturn layout.