Support "button label" hint for 8BitDo Pro Wired 2
Controllers such as 8BitDo Pro Wired 2 actually have their physical
labels according to the "Nintendo layout" (bottom = B, east = A) and
not the "Xbox layout" (default, bottom = A, east = B).

SDL supports the SDL_HINT_GAMECONTROLLER_USE_BUTTON_LABELS for this
purpose, which hints to the application that the user wants to follow
the labels in the controller, and not the physical layout of the
buttons [1]. That hint defaults to 1.

The mappings for this particular controller are made in the "default
layout" (hint=0) and do not support the "Nintendo layout" (hint=1),
which is inconsistent with how SDL works by default.

This commit addresses that by introducing a new copy of each one of the
three mappings for this controller (as it has 3 operations modes which
result in 3 different GUIDs [2]), which will activate depending on the
hint.

Naturally there are other controllers with the "Nintendo layout" that
would benefit from the same improvements, however the author of this
commit is restricting the scope of the change to the controller he
personally has and can test.

[1] https://github.com/libsdl-org/SDL/issues/6117
[2] https://github.com/gabomdq/SDL_GameControllerDB/pull/650