Use upstream mapping for 0079:0011 in Linux
lsusb:
```
ID 0079:0011 DragonRise Inc. Gamepad
```

I personally have this controller and the mappings from SDL upstream are
correct, while the mappings in the community DB are not.

This controller is likely to be a case where different manufacturers use
the same board (and VID/PID) but have different mappings due to
different firmware and/or wired differently.

This commit overwrites the community DB version with the upstream
version. While this can overwrite a good mapping for a different
controller, in case of conflicts such as this it's better if the
community DB agrees with upstream.

Joystick name used for CRC calculation: `USB Gamepad`
CRC from joystick name: `0xE86C`
GUID with CRC: `03006ce8790000001100000010010000`