Add final two Raphnet N64 adapters for Linux (#871)
* Add variant of the Raphnet N64 controller adapter that wasn't present before.

* Standardize name for Raphnet N64 Adapter addition.

* Remove CRC, correct platform field

* Additional Raphnet N64 Adapter added for Linux.

GUID given by SDL Gamepad Tool is  0300d8889b2800006400000001010000, but it seems to follow the pattern of the other Raphnet N64 Adapter added for Linux (seems same as Windows GUIDs for whole line of N64 Adapters besides the last four bytes).

* Add remaining two Raphnet N64 adapters to database for Linux

Not tested, but it should work when extrapolated from the other entries.

---------

Co-authored-by: Alynne <alviscera@gmail.com>